Finished! and added decimal to measurement

+++ src/hardware/Controller.h
@@ -46,7 +46,7 @@ public:
     void WiFiLoop();
     void reconnectWiFi();
     bool isWiFiConnected();
-    uint32_t getWeight();
+    float getWeight();
     bool isRTCConnected();
     ControllerState getState();
     void setState(ControllerState state);
@@ -55,7 +55,7 @@ public:
     void connectToWiFi(bool web_server, bool web_serial, bool OTA);
     void setUpWiFi(const char* ssid, const char* password, const char* hostname);
     bool hasIntervalPassed(uint32_t &previousMillis, uint32_t interval, bool to_min);
-    void broadcastWeight(uint32_t weight);
+    void broadcastWeight(float weight);
 
     void DEBUG_M(const char *message) {
         char buffer[100];
